Fixing linter so it doesn't provide the path to the command if not specified
In phpstan 0.11 and up the paths can (should) be specified in the phpstan config file.  If the path isn't specified in the .arclint file, no path will be appended to the command (otherwise phpstan will scan the entire directory since `./` is the default, which ends up being a lot more files than it should in some projects.